server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.2",
    "azure-storage-blob>=12.28.0",
    "celery[redis]>=5.6.2",
    "django>=6.0.2",
    "django-allauth[headless]>=65.14.1",
    "django-guardian>=3.2.0",
    "django-treebeard>=5.0.2",
    "djangorestframework>=3.16.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langsmith>=0.7.3",
    "llama-cloud>=1.4.0",
    "psycopg[binary,pool]>=3.3.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
]
